Fix warnings in engine

+++ Zend/zend_compile.h
@@ -470,7 +470,7 @@ struct _zend_op_array {
 	uint32_t T;         /* number of temporary variables */
 	/* END of common elements */
 
-	int cache_size;     /* number of run_time_cache_slots * sizeof(void*) */
+	size_t cache_size;  /* number of run_time_cache_slots * sizeof(void*) */
 	uint32_t last_var;  /* number of CV variables */
 	uint32_t last;      /* number of opcodes */
